Update default install location
Updated all install.sh scripts to install to conform to the XDG base
directory spec when installing for a single user. If no private bin
exists for the user, a private bin is created at ~/.local/bin

If ~/bin already exists, it is still preferred over ~/.local/bin. This
ensures that those that still use this path for their private bin have
that choice respected. If this directory doesn't exist, then we can
assume the user prefers the XDG spec for their private bin (or doesn't
care what the path of their private bin is)