Update genpass utility
Updated README.md

Changed genpass to make the charset a set before casting back to a list. This introduces some inherent randomness into the charset order as the data in a set is unordered. This will make it harder to crack passwords even if the underlying random number generator is flawed as the order of the charset used won't be as predictable.

Changed  to have default values for charset_size and iterations. A charset size is still required in order to set a number of iterations due to the use of sys.argv over something like argparse (may change this in the future)